backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
uuid
langgraph
langchain-groq
langchain-core
twilio
